add rabbitmq config resource
