+++ test/unit/mock/files/rabbitmq.config
@@ -1,0 +1,10 @@
+%% -*- mode: erlang -*-
+[
+ {rabbit,
+  [%% some comments...
+    {ssl_listeners, [5671]},
+    %% duplicate entries
+    {tcp_listeners, [5672]},
+    {tcp_listeners, [{"127.0.0.1", 5672},
+                     {"::1",       5672}]}
+]}].
